requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain Ecosystem
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.0

# Vector Database
chromadb==0.4.22

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2

# Embeddings & LLM
openai==1.54.0
tiktoken==0.8.0

# API Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# UI Framework
streamlit==1.29.0

# Utilities
requests==2.31.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
